backend/requirements.txt
aiohttp==3.10.10
elasticsearch==8.14.0
fastapi==0.115.3
fastapi_cache2==0.2.2
groq==0.11.0
httpx==0.27.2
jose==1.0.0
openai==1.52.1
pandas==2.2.3
pydantic==2.9.2
PyJWT==2.9.0
PyJWT==2.9.0
python-dotenv==1.0.1
python_jose==3.3.0
qdrant_client==1.12.0
regex==2024.7.24
Requests==2.32.3
tenacity==8.2.3
tiktoken==0.7.0
uvicorn==0.32.0
